langgraph_langflow/backend/requirements.txt
langgraph==1.0.10
langchain==1.2.10
openai==2.24.0
google-generativeai==0.8.6
python-dotenv==1.2.2
google-genai==1.65.0

fastapi 
uvicorn
uvicorn[standard]  # if not already installed
pydantic
fastapi-cors